// libs/shared/src/lib/models/courses/sp-courses/sp-course.model.ts
export interface ActiveSelfPacedCourse {
  id: string;
  createdAt: Date;
  updatedAt: Date;
  createdBy: string;
  updatedBy: string;
  agenda: any[];
  materials?: any;
  thumbnailKey: string;
  thumbnailUrl: string;
  hasCustomImage: boolean;
  bannerKey?: any;
  promotionMaterials: any[];
  promotionLinks: any[];
  code: string;
  sku: string;
  name: string;
  format: ConfigDto;
  keywords: string[];
  description: string;
  categoryName: string;
  categoryId: string;
  objectives?: any;
  targetAudience?: any;
  subcategoryId: string;
  subcategoryName: string;
  subcategoryCode: string;
  topic: any[];
  status?: any;
  maxSeats: number;
  minStudents: number;
  waitingListSize: number;
  waitingList: boolean;
  automaticApproval: boolean;
  availableForCourseCreation?: any;
  online?: any;
  level: ConfigDto;
  agendaEditable?: any;
  hierarchicalAgenda: any[];
  thumbnailProcess: boolean;
  thumbnailBucket: string;
  language: ConfigDto;
  editLock: boolean;
  domain: IDomainData;
  defaultSubDomain: string;
  parentThumbnailUrl?: any;
  brandedPortalMaterials?: any;
  parentId?: any;
  parentSku?: any;
  durationInHours?: any;
  durationInDays?: any;
  isInternal: boolean;
  pricing?: any;
  agoraProducts: any[];
  onlineDelivery?: any;
  registrationApprovalRequired?: any;
  categoryCode?: any;
  agoraProductSegmentCode?: any;
  agoraProductSegmentValue?: any;
  agoraProductSegmentNonEditable?: any;
  participationCertificate?: any;
  examCompletionAvailable: boolean;
  masterInternalRepositories: InternalRepositoryMaterial[];
  root: boolean;
  activeEventsCount?: any;
  totalEventsCount: number;
  archived?: any;
  deletable?: any;
  averageRating?: any;
  totalNumberOfReviews?: any;
  lxpEnabledDomain: boolean;
  publishStatus?: ConfigDto;
  lxpPrivate?: boolean;
  lxpRestrictUsers?: string[];
  lxpRestrictGroups?: { key: string; value: string }[];
  enrollmentPolicy?: ConfigDto;
  externalSKU?: string;
  specificExternalSKU?: boolean;
  lxpCourses?: PublishedCourseToLXP[];
}